Display comment function
I have added a comment function to this application. When users browse
an issue, they can click the comment button, and the app will use the
API to retrieve all comments associated with that issue. It's worth
noting that if the user doesn't click the comment button, the app will
not fetch the comments, ensuring greater efficiency.